[MIG] delivery_spring: Migration to 18.0
