+++ delivery_spring/README.rst
@@ -2,3 +2,9 @@ Delivery Spring
 ---------------
 
 Delivery Carrier connector for Spring
+
+TODO
+----
+
+- Abstract Spring API into it's own class
+
