+++ delivery_spring/models/__init__.py
@@ -1,1 +1,4 @@
-from . import delivery, stock
+from . import delivery_carrier
+from . import delivery_spring_service
+from . import stock_package_type
+from . import stock_picking
