+++ delivery_spring/models/delivery_spring_service.py
@@ -1,0 +1,10 @@
+from odoo import fields, models
+
+
+class SpringService(models.Model):
+    _name = "delivery.spring.service"
+    _description = "Spring Service"
+
+    active = fields.Boolean(default=True)
+    name = fields.Char()
+    ref = fields.Char()
