+++ delivery_spring/models/stock_package_type.py
@@ -1,13 +1,6 @@
 from odoo import fields, models
 
 
-class StockPicking(models.Model):
-    _inherit = "stock.picking"
-
-    spring_carrier = fields.Char()
-    spring_tracking_url = fields.Char()
-
-
 class StockPackageType(models.Model):
     _inherit = "stock.package.type"
