+++ delivery_state_events/__manifest__.py
@@ -13,5 +13,5 @@
         "data/cron.xml",
     ],
     "demo": [],
-    "license": "AGPL-3",
+    "license": "LGPL-3",
 }
